+++ saas/cloud-controller/pkg/kubernetes/templates/chart_template.go
@@ -2,8 +2,8 @@ package templates
 
 const (
 	ChartTemplate = `apiVersion: v2
-name: {{name}}
-description: {{name}} Applinth Implementation
+name: {{chartName}}
+description: {{displayName}} Applinth Implementation
 type: application
 version: 0.0.1
 appVersion: 0.0.1